backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
alembic==1.13.1
bcrypt==4.1.2
pymongo==4.6.1
motor==3.3.2
beanie==1.25.0
python-dotenv==1.0.0 